requirements.txt
Django==4.2
djangorestframework~=3.15.2
django-filter
celery~=5.4.0
redis
pandas~=2.2.3
boto3~=1.28.30